+++ src/main/java/bjack/venusrestblog/controller/UsersController.java
@@ -1,8 +1,11 @@
 package bjack.venusrestblog.controller;
 
+import bjack.venusrestblog.data.Post;
 import bjack.venusrestblog.data.User;
+import bjack.venusrestblog.misc.FieldHelper;
 import bjack.venusrestblog.repository.UsersRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -24,7 +27,11 @@ public class UsersController {
 
     @GetMapping("/{id}")
     public Optional<User> fetchUserById(@PathVariable long id) {
-        return usersRepository.findById(id);
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+        return optionalUser;
     }
 
     @GetMapping("/me")
@@ -49,6 +56,8 @@ public class UsersController {
 
     @PostMapping("/create")
     public void createUser(@RequestBody User newUser) {
+        // TODO: validate new user fields
+
         // don't need the below line at this point but just for kicks
         newUser.setCreatedAt(LocalDate.now());
         usersRepository.save(newUser);
@@ -56,22 +65,39 @@ public class UsersController {
 
     @DeleteMapping("/{id}")
     public void deleteUserById(@PathVariable long id) {
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
         usersRepository.deleteById(id);
     }
 
     @PutMapping("/{id}")
     public void updateUser(@RequestBody User updatedUser, @PathVariable long id) {
-        // find the post to update in the posts list
-        updatedUser.setId(id);
-        usersRepository.save(updatedUser);
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+        // get the user from the optional so we no longer have to deal with the optional
+        User originalUser = optionalUser.get();
+
+        // merge the changed data in updatedUser with originalUser
+        BeanUtils.copyProperties(updatedUser, originalUser, FieldHelper.getNullPropertyNames(updatedUser));
+
+        // originalUser now has the merged data (changes + original data)
+        originalUser.setId(id);
+
+        usersRepository.save(originalUser);
     }
 
     @PutMapping("/{id}/updatePassword")
     private void updatePassword(@PathVariable Long id, @RequestParam(required = false) String oldPassword, @RequestParam String newPassword) {
-        User user = usersRepository.findById(id).get();
-//        if(user == null) {
-//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User id " + id + " not found");
-//        }
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+
+        User user = optionalUser.get();
 
         // compare old password with saved pw
         if(!user.getPassword().equals(oldPassword)) {
